% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{causes}
\alias{causes}
\title{Causes of crashes.}
\format{A data frame with 888072 rows and 7 variables:
\describe{
  \item{id}{Unique number identifying the crash, by which the \code{link{crashes}} table can be joined.}
  \item{index}{Unique identifier of the cause of each crash (unique in combination with \code{id}.}
  \item{code}{Identifier of cause in the data dictionary.}
  \item{vehicle_id}{Vehicle in the crash with which the cause is associated.}
  \item{cause}{Description of the cause.}
  \item{cause_category}{Highest-level group to which the cause belongs.}
  \item{cause_subcategory}{Lowest-level group to which the cause belongs.}
}}
\source{
\url{http://www.nzta.govt.nz/resources/crash-analysis-system-data/}
}
\usage{
causes
}
\description{
A dataset of 888072 crash causes, recorded by the New Zealand Transport
Agency (NZTA) as being causes of crashes in the \code{\link{crashes}}
dataset.  There is a many-to-one relationship between causes and crashes.
They can be joined by the common column \code{id} (see the example below).
Alternatively, each cause can be associated with a specific vehicle in each
crash, by first joining the \code{\link{vehicles}} dataset to the
\code{\link{crashes}} one, and then joining the \code{causess} dataset by the
two common columns \code{id} and \code{vehicle_id}.

Other attributes in a many-to-one relationship with the crashes are
available in the datasets \code{\link{vehicles}}, and
\code{\link{objects_struck}}.
}
\seealso{
A comprehensive data dictionary is provided by the NZTA at
\url{http://www.nzta.govt.nz/assets/resources/guide-to-coded-crash-reports/docs/guide-to-coded-crash-reports.pdf}.
A copy of it is distributed with this package in the \code{extdata}
directory.

Details about how crash reports are completed are provided by the NZTA at
\url{http://www.nzta.govt.nz/assets/resources/traffic-crash-reports/docs/traffic-crash-reports.pdf}.
A copy of it is distributed with this package in the \code{extdata}
directory.
}
\keyword{datasets}

