% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{objects_struck}
\alias{objects_struck}
\title{Objects struck in crashes.}
\format{A data frame with 261276 rows and three variables:
\describe{
  \item{id}{Unique number identifying the crash, by which the \code{link{crashes}} table can be joined.}
  \item{object_order}{The order in which the object was struck in each crash (unique in combination with \code{id}.}
  \item{object}{The type object struck.}
}}
\source{
\url{http://www.nzta.govt.nz/resources/crash-analysis-system-data/}
}
\usage{
objects_struck
}
\description{
A dataset of 261276 objects, recorded by the New Zealand Transport
Agency (NZTA) as being struck during crashes in the \code{\link{crashes}}
dataset.  There is a many-to-one relationship between objects and crashes.
They can be joined by the common column \code{id} (see the example below).

Other attributes in a many-to-one relationship with the crashes are
available in the datasets \code{\link{vehicles}}, and
\code{\link{causes}}.
}
\seealso{
A comprehensive data dictionary is provided by the NZTA at
\url{http://www.nzta.govt.nz/assets/resources/guide-to-coded-crash-reports/docs/guide-to-coded-crash-reports.pdf}.
A copy of it is distributed with this package in the \code{extdata}
directory.

Details about how crash reports are completed are provided by the NZTA at
\url{http://www.nzta.govt.nz/assets/resources/traffic-crash-reports/docs/traffic-crash-reports.pdf}.
A copy of it is distributed with this package in the \code{extdata}
directory.
}
\keyword{datasets}

