% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicles}
\alias{vehicles}
\title{Vehicles involved in crashes.}
\format{A data frame with 979930 rows and three variables:
\describe{
  \item{id}{Unique number identifying the crash, by which the \code{link{crashes}} table can be joined.}
  \item{vehicle_id}{Unique identifier of the vehicle in each crash (unique in combination with \code{id}.}
  \item{vehicle}{Type of vehicle (car, truck, etc.).}
}}
\source{
\url{http://www.nzta.govt.nz/resources/crash-analysis-system-data/}
}
\usage{
vehicles
}
\description{
A dataset of 979930 vehicles, recorded by the New Zealand Transport Agency
(NZTA) as being involved in crashes in the \code{\link{crashes}} dataset.
There is a many-to-one relationship between vehicles and crashes.  They can
be joined by the common column \code{id} (see the example below).

Other attributes in a many-to-one relationship with the crashes are
available in the datasets \code{\link{causes}}, and
\code{\link{objects_struck}}.
}
\seealso{
A comprehensive data dictionary is provided by the NZTA at
\url{http://www.nzta.govt.nz/assets/resources/guide-to-coded-crash-reports/docs/guide-to-coded-crash-reports.pdf}.
A copy of it is distributed with this package in the \code{extdata}
directory.

Details about how crash reports are completed are provided by the NZTA at
\url{http://www.nzta.govt.nz/assets/resources/traffic-crash-reports/docs/traffic-crash-reports.pdf}.
A copy of it is distributed with this package in the \code{extdata}
directory.
}
\keyword{datasets}

